prov/verbs: Move RDM/RMA routines to the separate file
